discord-spider/requirements.txt
flask
gunicorn
pynacl